devtools/server/tests/chrome/test_inspector-scroll-into-view.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=901250
-->
<head>
  <meta charset="utf-8">
  <title>Test for Bug 901250</title>

  <script src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="chrome://mochikit/content/tests/SimpleTest/test.css">
  <script type="application/javascript" src="inspector-helpers.js"></script>
  <script type="application/javascript">
"use strict";

window.onload = async function() {
  SimpleTest.waitForExplicitFinish();

  const url = document.getElementById("inspectorContent").href;
  const { target, doc } = await attachURL(url);
  const inspector = await target.getFront("inspector");
  const walker = inspector.walker;

  const id = "#scroll-into-view";
  let rect = doc.querySelector(id).getBoundingClientRect();
  const nodeFront = await walker.querySelector(walker.rootNode, id);
  let inViewport = rect.x >= 0 &&
                   rect.y >= 0 &&
                   rect.y <= doc.defaultView.innerHeight &&
                   rect.x <= doc.defaultView.innerWidth;

  ok(!inViewport, "Element is not in viewport initially");

  await nodeFront.scrollIntoView();

  await new Promise(res => SimpleTest.executeSoon(res));

  rect = doc.querySelector(id).getBoundingClientRect();
  inViewport = rect.x >= 0 &&
                rect.y >= 0 &&
                rect.y <= doc.defaultView.innerHeight &&
                rect.x <= doc.defaultView.innerWidth;
  ok(inViewport, "Element is in viewport after calling nodeFront.scrollIntoView");

  await target.destroy();
  SimpleTest.finish();
};
  </script>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=901250">Mozilla Bug 901250</a>
<a id="inspectorContent" target="_blank" href="inspector-traversal-data.html">Test Document</a>
<p id="display"></p>
<div id="content" style="display: none">

</div>
<pre id="test">
</pre>
</body>
</html>
